requirements.txt
# Health Economic Modeling Hub - Dash Framework
# Requirements for modular Dash application

# Dash Framework & Components
dash==2.18.2
dash-bootstrap-components==1.6.0
plotly==5.24.1

# Data & Scientific Computing
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1

# Database
sqlalchemy==2.0.35

# AI & Agentic Framework
crewai==0.86.0
langgraph==0.2.62
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
crewai-tools==0.17.0

# Utilities
python-dotenv==1.0.1

# Development
pytest==8.3.3
black==24.10.0
